fix: file restructuring along with backend
