--- src/components/Question/questions.json
@@ -1,6 +1,0 @@
-[
-    {"question": "My question here", "answer": "Lorem ipsum dolor sit amet"},
-    {"question": "My question here", "answer": "Lorem ipsum dolor sit amet"}, 
-    {"question": "My question here", "answer": "Lorem ipsum dolor sit amet"}, 
-    {"question": "My question here", "answer": "Lorem ipsum dolor sit amet"} 
-]
