+++ src/pages/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Home from './Home';
+
+function App() {
+  return (
+    <div>
+      <Home />
+    </div>
+  );
+};
+
+export default App;
