+++ src/pages/Home.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import Testimonial from '../Testimonial/Testimonial';
+import Testimonial from '../components/Testimonial';
 
 // Data files to display
-import { testimonialData } from '../../data/testimonialData';
+import { testimonialData } from '../data/testimonialData';
 
 function Home() {
   return (
